move converter tests to the base class
