+++ flink/v1.20/flink/src/test/java/org/apache/iceberg/flink/source/TestIcebergSourceBoundedConverterBase.java
@@ -20,6 +20,7 @@ package org.apache.iceberg.flink.source;
 
 import java.io.IOException;
 import java.nio.file.Path;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -37,15 +38,20 @@ import org.apache.iceberg.Parameters;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.Table;
 import org.apache.iceberg.data.GenericAppenderHelper;
+import org.apache.iceberg.data.RandomGenericData;
 import org.apache.iceberg.data.Record;
 import org.apache.iceberg.expressions.Expression;
 import org.apache.iceberg.flink.FlinkConfigOptions;
 import org.apache.iceberg.flink.HadoopCatalogExtension;
 import org.apache.iceberg.flink.TableLoader;
 import org.apache.iceberg.flink.TestFixtures;
+import org.apache.iceberg.flink.TestHelpers;
 import org.apache.iceberg.flink.source.assigner.SimpleSplitAssignerFactory;
 import org.apache.iceberg.flink.source.reader.ReaderFunction;
 import org.apache.iceberg.relocated.com.google.common.collect.Lists;
+import org.apache.iceberg.relocated.com.google.common.collect.Sets;
+import org.apache.iceberg.types.TypeUtil;
+import org.junit.jupiter.api.TestTemplate;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.api.io.TempDir;
@@ -77,8 +83,37 @@ public abstract class TestIcebergSourceBoundedConverterBase<T> {
   @Parameter(index = 2)
   boolean useConverter;
 
-  static Table getUnpartitionedTable() {
-    return CATALOG_EXTENSION.catalog().createTable(TestFixtures.TABLE_IDENTIFIER, TestFixtures.SCHEMA);
+  @TestTemplate
+  public void testUnpartitionedTable() throws Exception {
+    Table table = CATALOG_EXTENSION.catalog().createTable(TestFixtures.TABLE_IDENTIFIER, TestFixtures.SCHEMA);
+    List<Record> expectedRecords = RandomGenericData.generate(TestFixtures.SCHEMA, 2, 0L);
+    new GenericAppenderHelper(table, fileFormat, temporaryFolder).appendToTable(expectedRecords);
+    TestHelpers.assertRecords(run(), expectedRecords, TestFixtures.SCHEMA);
+  }
+
+  @TestTemplate
+  public void testPartitionedTable() throws Exception {
+    String dateStr = "2020-03-20";
+    Table table = getPartitionedTable();
+    List<Record> expectedRecords = RandomGenericData.generate(TestFixtures.SCHEMA, 2, 0L);
+    for (Record expectedRecord : expectedRecords) {
+      expectedRecord.setField("dt", dateStr);
+    }
+    addRecordsToPartitionedTable(table, dateStr, expectedRecords);
+    TestHelpers.assertRecords(run(), expectedRecords, TestFixtures.SCHEMA);
+  }
+
+  @TestTemplate
+  public void testProjection() throws Exception {
+    Table table = getPartitionedTable();
+    List<Record> expectedRecords = RandomGenericData.generate(TestFixtures.SCHEMA, 2, 0L);
+    addRecordsToPartitionedTable(table, "2020-03-20", expectedRecords);
+    // select the "data" field (fieldId == 1)
+    Schema projectedSchema = TypeUtil.select(TestFixtures.SCHEMA, Sets.newHashSet(1));
+    List<Row> expectedRows =
+        Arrays.asList(Row.of(expectedRecords.get(0).get(0)), Row.of(expectedRecords.get(1).get(0)));
+    TestHelpers.assertRows(
+        run(projectedSchema, Collections.emptyList(), Collections.emptyMap()), expectedRows);
   }
 
   static Table getPartitionedTable() {
@@ -87,24 +122,20 @@ public abstract class TestIcebergSourceBoundedConverterBase<T> {
         .createTable(TestFixtures.TABLE_IDENTIFIER, TestFixtures.SCHEMA, TestFixtures.SPEC);
   }
 
-  protected TableLoader tableLoader() {
+  static TableLoader tableLoader() {
     return CATALOG_EXTENSION.tableLoader();
   }
 
-  protected void addRecordsToUnpartitionedTable(Table table, List<Record> expectedRecords) throws IOException {
-    new GenericAppenderHelper(table, fileFormat, temporaryFolder).appendToTable(expectedRecords);
-  }
-
-  protected void addRecordsToPartitionedTable(Table table, String dateStr, List<Record> expectedRecords) throws IOException {
+  private void addRecordsToPartitionedTable(Table table, String dateStr, List<Record> expectedRecords) throws IOException {
     new GenericAppenderHelper(table, fileFormat, temporaryFolder)
         .appendToTable(org.apache.iceberg.TestHelpers.Row.of(dateStr, 0), expectedRecords);
   }
 
-  List<Row> run() throws Exception {
+  private List<Row> run() throws Exception {
     return run(null, Collections.emptyList(), Collections.emptyMap());
   }
 
-  List<Row> run(
+  private List<Row> run(
       Schema projectedSchema, List<Expression> filters, Map<String, String> options)
       throws Exception {
 
@@ -160,7 +191,7 @@ public abstract class TestIcebergSourceBoundedConverterBase<T> {
         .flinkConfig(config);
   }
 
-  protected IcebergSource.Builder<T> createSourceBuilderWithReaderFunction(
+  private IcebergSource.Builder<T> createSourceBuilderWithReaderFunction(
       Table table, Schema projected, List<Expression> filters, Configuration config) throws Exception {
     return IcebergSource.<T>builder()
         .tableLoader(tableLoader())
