+++ flink/v1.20/flink/src/test/java/org/apache/iceberg/flink/source/TestIcebergSourceBoundedGenericRecord.java
@@ -18,8 +18,6 @@
  */
 package org.apache.iceberg.flink.source;
 
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import org.apache.avro.generic.GenericRecord;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
@@ -33,21 +31,15 @@ import org.apache.iceberg.Parameters;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.Table;
 import org.apache.iceberg.avro.AvroSchemaUtil;
-import org.apache.iceberg.data.RandomGenericData;
-import org.apache.iceberg.data.Record;
 import org.apache.iceberg.expressions.Expression;
 import org.apache.iceberg.flink.FlinkSchemaUtil;
 import org.apache.iceberg.flink.TestFixtures;
-import org.apache.iceberg.flink.TestHelpers;
 import org.apache.iceberg.flink.data.RowDataToRowMapper;
 import org.apache.iceberg.flink.sink.AvroGenericRecordToRowDataMapper;
 import org.apache.iceberg.flink.source.reader.AvroGenericRecordConverter;
 import org.apache.iceberg.flink.source.reader.AvroGenericRecordReaderFunction;
 import org.apache.iceberg.flink.source.reader.ReaderFunction;
 import org.apache.iceberg.flink.source.reader.RowDataConverter;
-import org.apache.iceberg.relocated.com.google.common.collect.Sets;
-import org.apache.iceberg.types.TypeUtil;
-import org.junit.jupiter.api.TestTemplate;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.apache.flink.configuration.Configuration;
 
@@ -62,39 +54,6 @@ public class TestIcebergSourceBoundedGenericRecord extends TestIcebergSourceBoun
       {FileFormat.PARQUET, 2, false},
       {FileFormat.ORC, 2, true}
     };
-  }
-
-  @TestTemplate
-  public void testUnpartitionedTable() throws Exception {
-    Table table = getUnpartitionedTable();
-    List<Record> expectedRecords = RandomGenericData.generate(TestFixtures.SCHEMA, 2, 0L);
-    addRecordsToUnpartitionedTable(table, expectedRecords);
-    TestHelpers.assertRecords(run(), expectedRecords, TestFixtures.SCHEMA);
-  }
-
-  @TestTemplate
-  public void testPartitionedTable() throws Exception {
-    String dateStr = "2020-03-20";
-    Table table = getPartitionedTable();
-    List<Record> expectedRecords = RandomGenericData.generate(TestFixtures.SCHEMA, 2, 0L);
-    for (Record expectedRecord : expectedRecords) {
-      expectedRecord.setField("dt", dateStr);
-    }
-    addRecordsToPartitionedTable(table, dateStr, expectedRecords);
-    TestHelpers.assertRecords(run(), expectedRecords, TestFixtures.SCHEMA);
-  }
-
-  @TestTemplate
-  public void testProjection() throws Exception {
-    Table table = getPartitionedTable();
-    List<Record> expectedRecords = RandomGenericData.generate(TestFixtures.SCHEMA, 2, 0L);
-    addRecordsToPartitionedTable(table, "2020-03-20", expectedRecords);
-    // select the "data" field (fieldId == 1)
-    Schema projectedSchema = TypeUtil.select(TestFixtures.SCHEMA, Sets.newHashSet(1));
-    List<Row> expectedRows =
-        Arrays.asList(Row.of(expectedRecords.get(0).get(0)), Row.of(expectedRecords.get(1).get(0)));
-    TestHelpers.assertRows(
-        run(projectedSchema, Collections.emptyList(), Collections.emptyMap()), expectedRows);
   }
 
   @Override
