+++ flink/v1.20/flink/src/test/java/org/apache/iceberg/flink/source/TestIcebergSourceBoundedRow.java
@@ -18,9 +18,6 @@
  */
 package org.apache.iceberg.flink.source;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.java.typeutils.RowTypeInfo;
 import org.apache.flink.streaming.api.datastream.DataStream;
@@ -29,57 +26,16 @@ import org.apache.flink.types.Row;
 import org.apache.iceberg.ParameterizedTestExtension;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.Table;
-import org.apache.iceberg.data.RandomGenericData;
-import org.apache.iceberg.data.Record;
 import org.apache.iceberg.flink.FlinkSchemaUtil;
-import org.apache.iceberg.flink.TestFixtures;
-import org.apache.iceberg.flink.TestHelpers;
 import org.apache.iceberg.flink.source.reader.RowConverter;
 import org.apache.iceberg.flink.source.reader.RowDataConverter;
-import org.apache.iceberg.relocated.com.google.common.collect.Sets;
-import org.apache.iceberg.types.TypeUtil;
-import org.junit.jupiter.api.TestTemplate;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 @ExtendWith(ParameterizedTestExtension.class)
 public class TestIcebergSourceBoundedRow extends TestIcebergSourceBoundedConverterBase<Row> {
 
-  @TestTemplate
-  public void testUnpartitionedTable() throws Exception {
-    Table table = getUnpartitionedTable();
-    List<Record> expectedRecords = RandomGenericData.generate(TestFixtures.SCHEMA, 2, 0L);
-    addRecordsToUnpartitionedTable(table, expectedRecords);
-    TestHelpers.assertRecords(run(), expectedRecords, TestFixtures.SCHEMA);
-  }
-
-  @TestTemplate
-  public void testPartitionedTable() throws Exception {
-    String dateStr = "2020-03-20";
-    Table table = getPartitionedTable();
-    List<Record> expectedRecords = RandomGenericData.generate(TestFixtures.SCHEMA, 2, 0L);
-    for (Record expectedRecord : expectedRecords) {
-      expectedRecord.setField("dt", dateStr);
-    }
-    addRecordsToPartitionedTable(table, dateStr, expectedRecords);
-    TestHelpers.assertRecords(run(), expectedRecords, TestFixtures.SCHEMA);
-  }
-
-  @TestTemplate
-  public void testProjection() throws Exception {
-    Table table = getPartitionedTable();
-    List<Record> expectedRecords = RandomGenericData.generate(TestFixtures.SCHEMA, 2, 0L);
-    addRecordsToPartitionedTable(table, "2020-03-20", expectedRecords);
-    // select the "data" field (fieldId == 1)
-    Schema projectedSchema = TypeUtil.select(TestFixtures.SCHEMA, Sets.newHashSet(1));
-    List<Row> expectedRows =
-        Arrays.asList(Row.of(expectedRecords.get(0).get(0)), Row.of(expectedRecords.get(1).get(0)));
-    TestHelpers.assertRows(
-        run(projectedSchema, Collections.emptyList(), Collections.emptyMap()), expectedRows);
-  }
-
-
   @Override
-  protected RowDataConverter<Row> getConverter(Schema icebergSchema, Table table) throws Exception {
+  protected RowDataConverter<Row> getConverter(Schema icebergSchema, Table table) {
     return RowConverter.fromIcebergSchema(icebergSchema);
   }
